+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 ### Added
 
 - MKS GEN L V1 boardmap based on ramps 1.4 (#175)
+- auto status report option to send periodic status info without ? required (#179)
 
 ### Changed